requirements.txt
crewai==0.95.0
crewai[tools]
fastapi
uvicorn[standard]
pandas
numpy
python-dotenv
openai
pydantic
requests
redis[asyncio]
